requirements.txt
Django==1.6.5
Fabric==1.9.0
Pillow==2.7.0
South==1.0
Werkzeug==0.9.6
XlsxWriter==0.7.3
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
cov-core==1.13.0
coverage==3.7.1
data-importer==2.2.0
django-celery==3.1.16
django-configurations==0.8
django-cors-headers==1.0.0
django-debug-panel==0.8.0
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-filter==0.11.0
django-filters==0.2.1
django-hosts==0.6
django-mailviews==0.6.3
django-model-utils==2.2
django-nose2==0.1.1
django-preserialize==1.1.0
-e git+git@github.com:Rustem/django-tastypie.git@a9d31f98610d7c7422786bdc20561c214e47068b#egg=django_tastypie-dev
django-tastypie-swagger==0.1.2
django-timezone-field==1.1
django-vcard==0.1
djangorestframework==3.2.4
djrill==1.1.1
ecdsa==0.11
epydoc==3.0.1
factory-boy==2.6.0
fake-factory==0.5.3
futures==3.0.3
google-api-python-client==1.4.0
httplib2==0.9
ipython==2.3.0
jdcal==1.2
kombu==3.0.26
mock==1.0.1
nose2==0.4.7
nose2-cov==1.0a4
oauth2client==1.4.6
openpyxl==2.1.4
paramiko==1.14.0
pip-autoremove==0.6.0
psycopg2==2.6.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pylibmc==1.3.0
python-dateutil==2.2
python-gflags==2.0
python-mimeparse==0.1.4
python-swiftclient==2.4.0
pytz==2014.7
requests==2.3.0
rsa==3.1.4
selenium==2.42.1
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.14
tldextract==1.5.1
uritemplate==0.6
vobject==0.8.1rc0
wsgiref==0.1.2
xlrd==0.9.3
django-almastorage==0.4
